Split receipt.ts file into receipt.ts, receiptTrie.ts and receiptProof.ts
