+++ ethereum_history_api/oracles/src/ethereum/receipt.test.ts
@@ -1,18 +1,8 @@
 import { describe, expect, it } from 'vitest';
-import {
-  ReceiptTrie,
-  encodeReceipt,
-  getReceiptProof,
-  getReceiptTrie,
-  logToRlpFields,
-  receiptToRlpFields,
-  txTypeToHex
-} from './receipt.js';
+import { encodeReceipt, logToRlpFields, receiptToRlpFields, txTypeToHex } from './receipt.js';
 import { toEventSelector } from 'viem';
 import { assert } from '../util/assert.js';
 import { loadReceiptFixture } from '../fixtures.js';
-import { createDefaultClient } from './client.js';
-import { stringify } from 'querystring';
 
 describe('logToRlpFields', () => {
   it(`log`, async () => {
@@ -167,20 +157,5 @@ describe('encodeReceipt', () => {
     const encodedReceipt = encodeReceipt(eip4344Receipt);
 
     expect(encodedReceipt.startsWith('0x03')).toBeTruthy();
-  });
-});
-
-describe('getReceiptProof', () => {
-  it('getReceiptProof success', async () => {
-    const BLOCK_NUMBER = 19411701n;
-    const client = createDefaultClient();
-    const receiptProof = await getReceiptProof(client, BLOCK_NUMBER, 0);
-
-    expect(receiptProof).toMatchInlineSnapshot(`
-      [
-        "0xf871a0f3479e13c3dec522fd4c32615ad70df480edba510cdda8d9bdb12ae5e607aba1a0c3bdc3854587e05e0355bf4b17e09fc5e3923783888a94b4965f7f7fe54faa68808080808080a0d3166b667cf8c0a77a4da7fe750aa40db5717963c50b2fc6ad04a7d8463d12308080808080808080",
-        "0xf904c130b904bd02f904b9018302e19ab9010000000000000000000000000000000000000000000000000000000000000004000000000000008000000000000000000002000001080000000000040000200800000000000000080008004008000000000000000000040000000000008000000000000000000000000000000000000000000000000000000010000050000000000000000001000000000000000000000000000000010000000000000000000000020000000000200000100000000004000000000000000000000000000000000000000402000000000000000000000000000000000400000000000800000000400010200000800000020400000000000000000000000000000000000000000000f903aef89b941111111254eeb25477b68fb85ed929f73a960582f842a0b9ed0243fdf00f0545c63a0af8850c090d86bb46682baec4bf3c496814fe4f02a0000000000000000000000000777999be819ffecee44a995560a9d0e97780a30cb8407a45f937345b3587c98d30c19cc470dad19e9f2db657691bad8e9a4e81a5fa360000000000000000000000000000000000000000000000000000000000000000f89b94c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000777999be819ffecee44a995560a9d0e97780a30ca00000000000000000000000008571c129f335832f6bbc76d49414ad2b8371a422a00000000000000000000000000000000000000000000000056bc75e2d63100000f89b94c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000008571c129f335832f6bbc76d49414ad2b8371a422a000000000000000000000000051c72848c68a965f66fa7a88855f9f7784502a7fa00000000000000000000000000000000000000000000000056bc75e2d63100000f89b94a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa000000000000000000000000051c72848c68a965f66fa7a88855f9f7784502a7fa0000000000000000000000000a88800cd213da5ae406ce248380802bd53b47647a00000000000000000000000000000000000000000000000000000005db3e3a688f89b94a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a0000000000000000000000000a88800cd213da5ae406ce248380802bd53b47647a00000000000000000000000001111111254eeb25477b68fb85ed929f73a960582a00000000000000000000000000000000000000000000000000000005db3e3a688f89b94a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000a88800cd213da5ae406ce248380802bd53b47647a0000000000000000000000000777999be819ffecee44a995560a9d0e97780a30ca00000000000000000000000000000000000000000000000000000005db3e3a688",
-      ]
-    `);
   });
 });
